Move the compression model classes
