+++ src/Files.App/Data/Contracts/ICompressArchiveModel.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) 2024 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
-namespace Files.App.Utils.Archives
+namespace Files.App.Data.Contracts
 {
 	/// <summary>
 	/// Represents an interface for archive creation support.
